Moved code into static method
