+++ src/BlockEntities/EnderChestEntity.cpp
@@ -8,7 +8,6 @@
 #include "../UI/EnderChestWindow.h"
 #include "../ClientHandle.h"
 #include "../Mobs/Ocelot.h"
-#include "../BoundingBox.h"
 
 
 
@@ -48,33 +47,15 @@ void cEnderChestEntity::SendTo(cClientHandle & a_Client)
 
 
 
-class cFindSittingCat :
-	public cEntityCallback
-{
-	virtual bool Item(cEntity * a_Entity) override
-	{
-		return (
-			(a_Entity->GetEntityType() == cEntity::etMonster) &&
-			(static_cast<cMonster *>(a_Entity)->GetMobType() == eMonsterType::mtOcelot) &&
-			(static_cast<cOcelot *>(a_Entity)->IsSitting())
-		);
-	}
-};
-
-
-
-
-
 bool cEnderChestEntity::UsedBy(cPlayer * a_Player)
 {
-	cFindSittingCat FindSittingCat;
 	if (
 		(GetPosY() < cChunkDef::Height - 1) &&
 		(
 			!cBlockInfo::IsTransparent(GetWorld()->GetBlock(GetPosX(), GetPosY() + 1, GetPosZ())) ||
 			(
 				(GetWorld()->GetBlock(GetPosX(), GetPosY() + 1, GetPosZ()) == E_BLOCK_AIR) &&
-				!GetWorld()->ForEachEntityInBox(cBoundingBox(Vector3d(GetPosX(), GetPosY() + 1, GetPosZ()), 1, 1), FindSittingCat)
+				!cOcelot::IsCatSittingOnBlock(GetWorld(), Vector3d(GetPos()))
 			)
 		)
 	)
